# Conflicts:
#	aioinflux/__init__.py
#	setup.py

+++ tests/conftest.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from aioinflux import InfluxDBClient
+from aioinflux import InfluxDBClient, testing_utils as utils
 
 with open(Path(__file__).parent / 'logging.yml') as f:
     logging.config.dictConfig(yaml.load(f))
@@ -19,25 +19,35 @@ def event_loop():
 
 
 @pytest.fixture(scope='module')
-def sync_client():
-    with InfluxDBClient(db='mytestdb', mode='blocking') as client:
-        client.create_database()
-        yield client
-        client.drop_database()
-
-
-@pytest.mark.asyncio
-@pytest.fixture(scope='module')
 async def async_client():
-    async with InfluxDBClient(db='mytestdb', mode='async') as client:
+    async with InfluxDBClient(db='async_client_test', mode='async') as client:
         await client.create_database()
         yield client
         await client.drop_database()
 
 
 @pytest.fixture(scope='module')
-def df_client():
-    with InfluxDBClient(db='mytestdb', mode='dataframe') as client:
+def sync_client():
+    with InfluxDBClient(db='sync_client_test', mode='blocking') as client:
         client.create_database()
         yield client
         client.drop_database()
+
+
+@pytest.fixture(scope='module')
+def df_client():
+    if utils.pd is None:
+        return
+    with InfluxDBClient(db='df_client_test', mode='blocking', output='dataframe') as client:
+        client.create_database()
+        yield client
+        client.drop_database()
+
+
+@pytest.fixture(scope='module')
+async def iter_client():
+    async with InfluxDBClient(db='iter_client_test', mode='async', output='iterable') as client:
+        await client.create_database()
+        await client.write([p for p in utils.cpu_load_generator(100)])
+        yield client
+        await client.drop_database()
